TIP-842: Move attribute option integration test to Structure
