+++ src/Akeneo/Pim/Structure/Component/tests/Integration/ExternalApi/AttributeOptionIntegration.php
@@ -1,15 +1,15 @@
 <?php
 
-namespace tests\integration\Pim\Component\Catalog\Normalizer\Standard;
+namespace Akeneo\Pim\Structure\Component\tests\Integration\Normalizer\ExternalApi;
 
-use Pim\Component\Catalog\tests\integration\Normalizer\Standard\AbstractStandardNormalizerTestCase;
+use Akeneo\Test\Integration\TestCase;
 
 /**
  * @author    Marie Bochu <marie.bochu@akeneo.com>
- * @copyright 2016 Akeneo SAS (http://www.akeneo.com)
+ * @copyright 2017 Akeneo SAS (http://www.akeneo.com)
  * @license   http://opensource.org/licenses/osl-3.0.php  Open Software License (OSL 3.0)
  */
-class AttributeOptionIntegration extends AbstractStandardNormalizerTestCase
+class AttributeOptionIntegration extends TestCase
 {
     public function testAttributeOption()
     {
@@ -25,8 +25,16 @@ class AttributeOptionIntegration extends AbstractStandardNormalizerTestCase
         $repository = $this->get('pim_catalog.repository.attribute_option');
         $serializer = $this->get('pim_serializer');
 
-        $result = $serializer->normalize($repository->findOneByIdentifier('a_multi_select.optionA'), 'standard');
+        $result = $serializer->normalize($repository->findOneByIdentifier('a_multi_select.optionA'), 'external_api');
 
         $this->assertSame($expected, $result);
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    protected function getConfiguration()
+    {
+        return $this->catalog->useTechnicalCatalog();
     }
 }
