+++ src/Akeneo/Pim/Structure/Component/tests/Integration/Normalizer/Standard/AttributeOptionIntegration.php
@@ -1,13 +1,15 @@
 <?php
 
-namespace Akeneo\Tool\Component\Api\tests\integration\Normalizer;
+namespace Akeneo\Pim\Structure\Component\tests\Integration\Normalizer\Standard;
+
+use Akeneo\Test\Integration\TestCase;
 
 /**
  * @author    Marie Bochu <marie.bochu@akeneo.com>
- * @copyright 2017 Akeneo SAS (http://www.akeneo.com)
+ * @copyright 2016 Akeneo SAS (http://www.akeneo.com)
  * @license   http://opensource.org/licenses/osl-3.0.php  Open Software License (OSL 3.0)
  */
-class AttributeOptionIntegration extends AbstractNormalizerTestCase
+class AttributeOptionIntegration extends TestCase
 {
     public function testAttributeOption()
     {
@@ -23,8 +25,16 @@ class AttributeOptionIntegration extends AbstractNormalizerTestCase
         $repository = $this->get('pim_catalog.repository.attribute_option');
         $serializer = $this->get('pim_serializer');
 
-        $result = $serializer->normalize($repository->findOneByIdentifier('a_multi_select.optionA'), 'external_api');
+        $result = $serializer->normalize($repository->findOneByIdentifier('a_multi_select.optionA'), 'standard');
 
         $this->assertSame($expected, $result);
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    protected function getConfiguration()
+    {
+        return $this->catalog->useTechnicalCatalog();
     }
 }
